requirements.txt
openai>=1.40.0
langgraph>=1.0.0
langchain-core>=0.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
